feat: added request body in Payload tab
