+++ browser-extension/common/src/devtools/containers/network/details-tabs/PayloadTabContent/components/RequestBody/RequestBody.tsx
@@ -1,0 +1,34 @@
+import React from "react";
+
+import { NetworkEvent } from "../../../../../../types";
+import { ObjectInspector } from "@devtools-ds/object-inspector";
+
+interface Props {
+  networkEvent: NetworkEvent;
+  parsed?: Boolean;
+}
+
+const RequestBodyParsed: React.FC<Props> = ({ networkEvent }) => {
+  return (
+    <ObjectInspector
+      expandLevel={3}
+      data={JSON.parse(networkEvent.request?.postData?.text)}
+      includePrototypes={false}
+      className="object-inspector"
+    />
+  );
+};
+
+const RequestBodySource: React.FC<Props> = ({ networkEvent }) => {
+  return <div className="request-body-content">{networkEvent.request?.postData?.text}</div>;
+};
+
+const RequestBody: React.FC<Props> = ({ networkEvent, parsed }) => {
+  if (parsed) {
+    return <RequestBodyParsed networkEvent={networkEvent} />;
+  }
+
+  return <RequestBodySource networkEvent={networkEvent} />;
+};
+
+export default RequestBody;
