+++ browser-extension/common/src/devtools/containers/network/details-tabs/PayloadTabContent/components/RequestBody/RequestBodyParsed.tsx
@@ -1,0 +1,25 @@
+import React from "react";
+
+import { NetworkEvent } from "../../../../../../types";
+import { ObjectInspector } from "@devtools-ds/object-inspector";
+
+interface Props {
+  networkEvent: NetworkEvent;
+  parsed?: Boolean;
+}
+
+// Displaying on JSON parsed body for now
+// Future Improvements: Form-Data (application/x-www-form-urlencoded)
+const RequestBodyParsed: React.FC<Props> = ({ networkEvent }) => {
+  let parsedBody = networkEvent.request?.postData?.text;
+
+  try {
+    parsedBody = JSON.parse(networkEvent.request?.postData?.text);
+  } catch (err) {
+    console.log("Unable to parse Request Body");
+  }
+
+  return <ObjectInspector expandLevel={3} data={parsedBody} includePrototypes={false} className="object-inspector" />;
+};
+
+export default RequestBodyParsed;
