+++ browser-extension/common/src/devtools/utils.ts
@@ -109,3 +109,11 @@ export const matchResourceTypeFilter = (
       return true;
   }
 };
+
+export const isRequestBodyParseable = (mimeType: string): boolean => {
+  if (mimeType && mimeType.startsWith("application/json")) {
+    return true;
+  }
+
+  return false;
+};
